add repositories and commands

+++ database/migrations/2017_02_09_154054_create_days_table.php
@@ -14,7 +14,7 @@ class CreateDaysTable extends Migration
     public function up()
     {
         Schema::create('days', function (Blueprint $table) {
-            $table->increments('id');
+            $table->tinyInteger('id');
             $table->string("name", 20);
         });
     }
